refactor: Split reqsign-azure-storage (#474)
