+++ crates/reqsign-azure-storage/Cargo.toml
@@ -1,0 +1,40 @@
+[package]
+name = "reqsign-azure-storage"
+version = "0.1.0"
+
+categories.workspace = true
+description.workspace = true
+documentation.workspace = true
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+
+[dependencies]
+anyhow.workspace = true
+chrono.workspace = true
+form_urlencoded.workspace = true
+http.workspace = true
+log.workspace = true
+percent-encoding.workspace = true
+reqsign.workspace = true
+reqwest.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+
+
+[dev-dependencies]
+aws-credential-types = "1.1.8"
+aws-sigv4 = "1.2.0"
+criterion.workspace = true
+dotenv.workspace = true
+env_logger.workspace = true
+hex.workspace = true
+macro_rules_attribute.workspace = true
+once_cell.workspace = true
+pretty_assertions.workspace = true
+reqwest = { workspace = true, features = ["rustls-tls"] }
+sha2.workspace = true
+temp-env.workspace = true
+tempfile.workspace = true
+test-case.workspace = true
+tokio = { workspace = true, features = ["full"] }
