+++ crates/reqsign-azure-storage/src/constants.rs
