+++ crates/reqsign-azure-storage/src/workload_identity_credential.rs
