--- crates/reqsign/src/azure/mod.rs
@@ -1,7 +1,0 @@
-//! Azure Storage SharedKey support
-//!
-//! Use [`azure::storage::Signer`][crate::azure::storage::Signer]
-mod storage;
-pub use storage::*;
-
-mod constants;
