--- crates/reqsign/src/azure/storage/sas/mod.rs
@@ -1,1 +1,0 @@
-pub mod account_sas;
