--- crates/reqsign/tests/azure/mod.rs
@@ -1,1 +1,0 @@
-mod storage;
